Move test utils to proper locations.
